sofacomerda

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: anolivei <anolivei@student.42sp.org>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/29 22:48:50 by anolivei          #+#    #+#             */
-/*   Updated: 2020/03/14 04:53:47 by anolivei         ###   ########.fr       */
+/*   Updated: 2020/03/15 02:42:44 by anolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ int		ft_print_hexa(va_list arguments, char *lower_upper,  t_flags flag);
 int		ft_print_pointer(va_list arguments, t_flags flag);
 int		ft_print_u(va_list arguments, t_flags flag);
 int		ft_print_percent(t_flags flag);
-size_t	ft_strlen(const char *str);
+int		ft_strlen(const char *str);
 int		ft_putchar(char c);
 int		ft_putstr(char *c);
